Move GeneralTemplate to general package to match convention
